BUILD: http-htx: fix build warning regarding long type in printf
Commit a66adf41e ("MINOR: http-htx: Add understandable errors for the
errorfiles parsing") added a warning when loading malformed error files,
but this warning may trigger another build warning due to the %lu format
used. Let's simply cast it for output since it's just used for end user
output.

This must be backported to 2.0 like the commit above.

(cherry picked from commit 431a12cafeeec7300b7cea7e19b892d4e8c4900d)
(cherry picked from commit 37b9630d0a93ff2ffc89a85a095e41db72de1c5b)
(cherry picked from commit 5d74de59f9428b2e1a3a8b58505bf9e7d187ea76)
(cherry picked from commit eaedb767a7b2fbe2de2603b46ecdd6a89034318f)

+++ src/http_htx.c
@@ -699,7 +699,7 @@ int http_str_to_htx(struct buffer *buf, struct ist raw, char **errmsg)
 
 		memprintf(errmsg, "payload size does not match the announced content-length (%lu != %lu)."
 			  " C-L header is updated accordingly but it should be fixed to avoid any errors on future versions.",
-			  (raw.len - ret), h1m.body_len);
+			  (unsigned long)(raw.len - ret), (unsigned long)h1m.body_len);
 
 		for (i = 0; hdrs[i].n.len; i++) {
 			if (isteqi(hdrs[i].n, ist("content-length")))
